* experiment/glfw:
  fixed a type conversion warning
  Implemented the input mechanism using GLFW's input callbacks.
  Implemented the GLFW buffer-swapping and event poling code. We can now run to the main nTA screen!!!
  fio_FindFirstFile() needs to support wildcard matching (/foo/*.bar). To support this on macOS we need to introduce a small helper class, _fio_FindOperation, that contains the find state and checks enumeration results with an NSPredicate that supports wildcard matching.
  Added the 'OTA' and 'root' distribution folders as base resources for the Cocoa nTA app. I had trouble getting the app to use Distribution as its working directory.
  Added the ability for app_Console to print() comments to stdout. This will help with debugging from an IDE without having to hunt down the log file.
  Initial changes to run nTA with GLFW + OpenGL.

+++ nTA/Gfx Dll Projects/OpenGL/Source/ogl_Model.cpp
@@ -723,7 +723,7 @@ BOOL ogl_Model::AssembleTexture( UsedTextureList_t& Textures )
 			if (NeedsPaletteConversion)
 			{
 				const PALETTEENTRY *palette = static_cast<PALETTEENTRY *>(pPalette);
-				for (int a = 0, b = 0, n = pImage->Size.width; a < n; ++a, b += TextureImage.Stride)
+				for (long a = 0, b = 0, n = pImage->Size.width; a < n; ++a, b += TextureImage.Stride)
 				{
 					const BYTE index = pImg[a];
 					const PALETTEENTRY *entry = palette + index;
